アンインストールの不具合修正

+++ apps/stirling/tools/chocolateyuninstall.ps1
@@ -5,7 +5,9 @@ $installPath = Join-Path (Get-ToolsLocation) $env:ChocolateyPackageName
 if (Test-Path $installPath) {
     # Remove Shortcut
     $lnkpath = Join-Path ([Environment]::GetFolderPath('Desktop')) ("Stirling.lnk")
-    Remove-Item -path $lnkpath -Force
+    if (Test-Path $lnkpath) {
+        Remove-Item -path $lnkpath -Force
+    }
 
     # Remove Install File
     Remove-Item -path $installPath -Recurse -Force
